move center() function from patch3 to element3
